BACKPORT: ASoC: SOF: Intel: Constify sof_probe_compressed_ops
The only usage of sof_probe_compressed_ops is to assign its address to
the compress_ops field in the snd_soc_component_driver struct, which is
a pointer to const. The assignment is done in sound/soc/sof/pcm.c. Make
it const to allow the compiler to put it in read-only memory.

Signed-off-by: Rikard Falkeborn <rikard.falkeborn@gmail.com>
Link: https://lore.kernel.org/r/20210526231013.46530-6-rikard.falkeborn@gmail.com
Signed-off-by: Mark Brown <broonie@kernel.org>

BUG=b:238403794
TEST=Test Audio use cases

Conflicts: sound/soc/sof/compress.c
	sound/soc/sof/compress.h
(conflict resolved by deleting two files)
Signed-off-by: sboyane <shankarx.boyane@intel.com>
Change-Id: If6095518401c8f216e988e13e2ce5de2e10cb8fd